% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoHiM.R
\name{CoHiM}
\alias{CoHiM}
\title{Run CoHiM for null hypotheses testing under dependence}
\usage{
CoHiM(p_matrix, q = 0.05)
}
\arguments{
\item{p_matrix}{a n by J matrix of p-values from n studies for J SNPs}

\item{q}{FDR nominal}
}
\value{
List containing
\describe{
\item{rLIS_mat}{Matrix of pairwise replicability scores}
\item{pi_mat}{Matrix of estimated stationary probabilities}
\item{A_list}{List of transition matrices}
\item{f1_mat}{Matrix of estimated alternative densities (study A)}
\item{f2_mat}{Matrix of estimated alternative densities (study B)}
\item{e_value}{Vector of e-values}
\item{q_star}{Optimal FDR threshold level}
\item{rej_index}{Logical vector of discovery decisions}
}
}
\description{
Run CoHiM for null hypotheses testing under dependence
}
