% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_eBH.R
\name{min_eBH}
\alias{min_eBH}
\title{Calculatet e-values and conduct eBH procedure for discoveries}
\usage{
min_eBH(rLIS_mat, null_prop_list, q = 0.05)
}
\arguments{
\item{rLIS_mat}{rLIS matrix with n_pair rows and m columns}

\item{null_prop_list}{null proportion estimation for each pair}

\item{q}{FDR nominal}
}
\value{
A list containing:
\item{q_star}{Optimal FDR threshold level}
\item{rej_num}{Number of rejected hypotheses}
\item{e_value}{Vector of e-values for each hypothesis}
\item{rej_index}{Logical vector indicating rejection decisions}
}
\description{
Calculatet e-values and conduct eBH procedure for discoveries
}
